+++ docs/conf.py
@@ -70,6 +70,7 @@ intersphinx_mapping = {
     "dynaconf": ("https://dynaconf.readthedocs.io/en/latest/", None),
     "jwt": ("https://pyjwt.readthedocs.io/en/latest/", None),
     "python": ("https://docs.python.org/3/", None),
+    "structlog": ("http://www.structlog.org/en/stable/", None),
     "wtforms": ("https://wtforms.readthedocs.io/en/stable/", None),
 }