+++ src/gafaelfawr/handlers/login.py
@@ -15,7 +15,7 @@ from gafaelfawr.providers.base import ProviderException
 
 if TYPE_CHECKING:
     from gafaelfawr.factory import ComponentFactory
-    from logging import Logger
+    from structlog import BoundLogger
 
 __all__ = ["get_login"]
 
@@ -80,7 +80,7 @@ async def redirect_to_provider(request: web.Request) -> web.Response:
         external authentication provider.
     """
     factory: ComponentFactory = request.config_dict["gafaelfawr/factory"]
-    logger: Logger = request["safir/logger"]
+    logger: BoundLogger = request["safir/logger"]
 
     # Determine where the user is trying to go.
     session = await get_session(request)
@@ -94,6 +94,7 @@ async def redirect_to_provider(request: web.Request) -> web.Response:
         msg = "No destination URL specified"
         logger.warning(msg)
         raise web.HTTPBadRequest(reason=msg, text=msg)
+    logger = logger.bind(return_url=return_url)
     if urlparse(return_url).hostname != request.url.raw_host:
         msg = f"Redirect URL not at {request.host}"
         logger.warning(msg)
@@ -132,7 +133,7 @@ async def redirect_to_provider(request: web.Request) -> web.Response:
         session["state"] = state
 
     # Get the authentication provider URL send the user there.
-    auth_provider = factory.create_provider(request)
+    auth_provider = factory.create_provider(request, logger)
     redirect_url = auth_provider.get_redirect_url(state)
     raise web.HTTPSeeOther(redirect_url)
 
@@ -160,7 +161,7 @@ async def handle_provider_return(request: web.Request) -> web.Response:
         external authentication provider.
     """
     factory: ComponentFactory = request.config_dict["gafaelfawr/factory"]
-    logger: Logger = request["safir/logger"]
+    logger: BoundLogger = request["safir/logger"]
 
     # Extract details from the reply, check state, and get the return URL.
     session = await get_session(request)
@@ -170,22 +171,28 @@ async def handle_provider_return(request: web.Request) -> web.Response:
         msg = "Authentication state mismatch"
         raise web.HTTPForbidden(reason=msg, text=msg)
     return_url = session.pop("rd")
+    logger = logger.bind(return_url=return_url)
 
     # Build a session based on the reply from the authentication provider.
-    auth_provider = factory.create_provider(request)
+    auth_provider = factory.create_provider(request, logger)
     try:
         auth_session = await auth_provider.create_session(code, state)
     except ProviderException as e:
-        logger.warning("Provider authentication failed: %s", str(e))
+        logger.warning("Provider authentication failed", error=str(e))
         raise web.HTTPInternalServerError(reason=str(e), text=str(e))
-    except ClientResponseError:
+    except ClientResponseError as e:
         msg = "Cannot contact authentication provider"
-        logger.exception(msg)
+        logger.exception(msg, error=str(e))
         raise web.HTTPInternalServerError(reason=msg, text=msg)
 
     # Store the session and send the user back to what they were doing.
     session = await new_session(request)
     session["handle"] = auth_session.handle.encode()
+    logger = logger.bind(
+        user=auth_session.token.username,
+        token=auth_session.token.jti,
+        scope=" ".join(sorted(auth_session.token.scope)),
+    )
     logger.info(
         "Successfully authenticated user %s (%s)",
         auth_session.token.username,