+++ src/gafaelfawr/handlers/well_known.py
@@ -10,6 +10,7 @@ from gafaelfawr.handlers import routes
 
 if TYPE_CHECKING:
     from gafaelfawr.config import Config
+    from structlog import BoundLogger
 
 __all__ = ["get_well_known_jwks"]
 
@@ -31,6 +32,8 @@ async def get_well_known_jwks(request: web.Request) -> web.Response:
         The outgoing response.
     """
     config: Config = request.config_dict["gafaelfawr/config"]
+    logger: BoundLogger = request["safir/logger"]
 
     jwks = config.issuer.keypair.public_key_as_jwks(kid=config.issuer.kid)
+    logger.info("Returned JWKS")
     return web.json_response({"keys": [jwks]})