+++ tests/handlers/logout_test.py
@@ -2,14 +2,19 @@
 
 from __future__ import annotations
 
+import json
 from typing import TYPE_CHECKING
+from unittest.mock import ANY
 from urllib.parse import parse_qs, urlparse
 
 if TYPE_CHECKING:
+    from _pytest.logging import LogCaptureFixture
     from tests.setup import SetupTestCallable
 
 
-async def test_logout(create_test_setup: SetupTestCallable) -> None:
+async def test_logout(
+    create_test_setup: SetupTestCallable, caplog: LogCaptureFixture
+) -> None:
     setup = await create_test_setup("github")
 
     # Simulate the initial authentication request.
@@ -36,9 +41,21 @@ async def test_logout(create_test_setup: SetupTestCallable) -> None:
     assert r.status == 200
 
     # Go to /logout without specifying a redirect URL and check the redirect.
+    caplog.clear()
     r = await setup.client.get("/logout", allow_redirects=False)
     assert r.status == 303
     assert r.headers["Location"] == setup.config.after_logout_url
+    data = json.loads(caplog.record_tuples[0][2])
+    assert data == {
+        "event": "Successful logout",
+        "level": "info",
+        "logger": "gafaelfawr",
+        "method": "GET",
+        "path": "/logout",
+        "remote": "127.0.0.1",
+        "request_id": ANY,
+        "user_agent": ANY,
+    }
 
     # Confirm that we're no longer logged in.
     r = await setup.client.get("/auth", params={"scope": "read:all"})
@@ -85,13 +102,25 @@ async def test_logout_with_url(create_test_setup: SetupTestCallable) -> None:
 
 
 async def test_logout_not_logged_in(
-    create_test_setup: SetupTestCallable,
+    create_test_setup: SetupTestCallable, caplog: LogCaptureFixture
 ) -> None:
     setup = await create_test_setup()
 
     r = await setup.client.get("/logout", allow_redirects=False)
     assert r.status == 303
     assert r.headers["Location"] == setup.config.after_logout_url
+    data = json.loads(caplog.record_tuples[0][2])
+    assert data == {
+        "event": "Logout of already-logged-out session",
+        "level": "info",
+        "logger": "gafaelfawr",
+        "method": "GET",
+        "path": "/logout",
+        "remote": "127.0.0.1",
+        "request_id": ANY,
+        "user_agent": ANY,
+    }
+
     r = await setup.client.get("/auth", params={"scope": "read:all"})
     assert r.status == 401